% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{calculate_type3_mi}
\alias{calculate_type3_mi}
\title{calculate_type3_mi}
\usage{
calculate_type3_mi(mira_obj, vcov_fun = NULL)
}
\arguments{
\item{mira_obj}{A \code{mira} object containing multiply imputed model fits.}

\item{vcov_fun}{Optional function used to calculate each model covariance matrix.}
}
\description{
calculate_type3_mi
}
\details{
The function calculates the  3 p-values based on Wald's statistics.
}
