% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cardio_data}
\alias{cardio_data}
\title{Cardiology Patient Dataset}
\format{
A data frame with 185 rows and 20 variables
}
\usage{
data(cardio_data)
}
\description{
A synthetic dataset of 185 patients who underwent cardiac surgery.
The variables are as follows:
}
\details{
\itemize{
  \item PatientID: Unique identifier for each patient
  \item Age: Age in years
  \item Sex: Biological sex
  \item NYHA: NYHA functional class
  \item EjectionFraction: Ejection fraction (%)
  \item BMI: Body mass index (kg/m²)
  \item HeartRate: Heart rate (bpm)
  \item SBP: Systolic blood pressure (mmHg)
  \item DBP: Diastolic blood pressure (mmHg)
  \item SmokingStatus: Smoking status: Never, Former, Current
  \item HistoryMI: History of myocardial infarction
  \item SurgeryDate: Date of cardiac surgery
  \item TransplantDate: Date of heart transplant
  \item DeathDate: Date of death
  \item LastVisitDate: Date of last clinical visit
  \item Diabetes: Presence of diabetes
  \item Hypertension: Presence of hypertension
  \item COPD: Presence of COPD
  \item CKD: Presence of chronic kidney disease
  \item CAD: Presence of coronary artery disease
}
}
\keyword{datasets}
