% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cardio_data_dictionary}
\alias{cardio_data_dictionary}
\title{Cardiology Patient Dataset dictionary}
\format{
A data frame with 20 rows and 3 columns:
}
\usage{
data(cardio_data_dictionary)
}
\description{
A synthetic dataset of 185 patients who underwent cardiac surgery.
The variables are as follows:
}
\details{
\itemize{
  \item VariableName: Name of the variable in the dataset
  \item Label: Manuscript-ready label
  \item Description: Detailed description of the variable
}
}
\keyword{datasets}
