% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_set_up.R
\name{check_box_convert}
\alias{check_box_convert}
\title{check box convert}
\usage{
check_box_convert(df, check_box_cols, title = NULL)
}
\arguments{
\item{df}{A data frame containing the checkbox-style variables.}

\item{check_box_cols}{Character vector of column names representing the checkbox-style question. The columns must be either logical or 0/1}

\item{title}{Optional character string used to label the checkbox group in the output table.}
}
\value{
A modified data frame with checkbox columns converted to logical type and missingness handled. If `title` is provided, an additional column is added to group the checkbox variables.
}
\description{
Converts checkbox-style variables into logical format for inclusion in a Table 1. Ensures that checkbox responses are treated as binary indicators and handles missingness appropriately. To be used with \code{\link{table_one}}
}
\details{
This function prepares checkbox-style questions (i.e., multiple binary columns representing selections) for descriptive analysis. It converts specified columns to logical type and sets all values to `NA` for rows where none of the checkbox options were selected. Optionally, it adds a title column to group these variables visually in the output table.
}
\examples{

library(dplyr)

Comorbidities  <- cardio_data \%>\% select(Diabetes:NoComorbidities) \%>\% names()

## Without checkbox question (Assuming missing all questions is FALSE)
cardio_data \%>\% select(Comorbidities) \%>\%
 table_one()
# Example with one checkbox question (Assumes missing all questions is missing)

cardio_data \%>\% select(Comorbidities) \%>\%
 check_box_convert(check_box_cols = Comorbidities,title = "Comorbidities¹")  \%>\%
 table_one(Check_box = Comorbidities,
           Check_box_title = "Comorbidities¹") \%>\%
 kable_table_one() \%>\%
 kableExtra::footnote(
   general = "¹Patients could present with more than one comorbidity, totals may not sum to 100\%.",
   general_title = "",
   footnote_as_chunk = TRUE
 )


## Example with two different checkbox questions

Comorbidities1  <- cardio_data \%>\% select(Diabetes:COPD) \%>\% names()
Comorbidities2  <- cardio_data \%>\% select(CKD:CAD) \%>\% names()



cardio_data \%>\% select(Comorbidities1,Comorbidities2) \%>\%
 check_box_convert(check_box_cols = Comorbidities1,title = "Comorbidities1¹")  \%>\%
 check_box_convert(check_box_cols = Comorbidities2,title = "Comorbidities2¹")  \%>\%
 table_one(Check_box = c(Comorbidities1, Comorbidities2),
           Check_box_title = c("Comorbidities1¹","Comorbidities2¹")) \%>\%
 kable_table_one() \%>\%
 kableExtra::footnote(
   general = "¹Patients could present with more than one comorbidity, totals may not sum to 100\%.",
   general_title = "",
   footnote_as_chunk = TRUE
 )


}
