% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{decimalplaces}
\alias{decimalplaces}
\title{decimalplaces}
\usage{
decimalplaces(x, max_dec = 4L)
}
\arguments{
\item{x}{a continuous variable}

\item{max_dec}{Maximum number of decimal places to return.}
}
\value{
the most frequent number of digits in the variable
}
\description{
decimalplaces
}
\details{
An internal function that determines the number of digits in the summary of a continuous variable.
}
