% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_functions.R
\name{estimate_cif_km}
\alias{estimate_cif_km}
\title{Kaplan-Meier survival and cumulative incidence (CIF) estimates}
\usage{
estimate_cif_km(df, evt_time, evt, group, conf.type = "default", ...)
}
\arguments{
\item{df}{A data frame containing the survival data.}

\item{evt_time}{A numeric vector representing the time to event or censoring.}

\item{evt}{For survival data: a numeric event indicator (1 = event occurred, 0 = censored).
For competing risks data: a factor event indicator (0 = censored, 1 = Event 1, 2 = Event 2, ...).}

\item{group}{A grouping variable for stratified survival curves.}

\item{conf.type}{The confidence-interval transformation passed to \code{survival::survfit()}. With the default value, Kaplan-Meier estimates use the log-log transformation and CIF estimates use the log transformation.}

\item{...}{Additional arguments passed to \code{survival::survfit()}.}
}
\value{
A \code{survfit} object containing the survival estimates.
}
\description{
Computes Kaplan-Meier survival estimates and cumulative incidence (CIF) from a dataset, optionally stratified by a grouping variable. #See also \code{\link{construct_surv_cmprisk_var}},\code{\link{summarize_km}}, \code{\link{summarize_cif}}, \code{\link{show_surv}}, and \code{\link{show_cif}}   for related functions.
}
\details{
The function will produce  Kaplan-Meier survival estimates when evt is numeric and will produce CIF estimates when evt is a factor

The function analyzes the data (df) using the Kaplan-Meier method with pointwise 95% confidence intervals. By default, the intervals use the log-log transformation. For competing-risks data, the default intervals use the log transformation. The input data are stored in the call(), which can be used in
run_logrank_test().

The function analyzes the competing data (df) using Andersen-Johansen method in estimating cumulative incidence
function.The function store the input data in the call(), which can be used in run_gray_test().
}
\examples{


## For survival

survival_data <-  construct_surv_cmprisk_var(cardio_data,
                                            patid = PatientID,
                                            idx_dt = SurgeryDate,
                                            evt_dt = DeathDate,
                                            end_dt = LastVisitDate,
                                            append = TRUE,
                                            units = "months",
                                            adm_cnr_time = 24)

KM <- estimate_cif_km(survival_data, evt = evt,evt_time = evt_time)
KM_Sex <- estimate_cif_km(survival_data, evt = evt,evt_time = evt_time,group = Sex)


# For competing risks
cmp_risk_data <- construct_surv_cmprisk_var(cardio_data,
                                           patid = PatientID,
                                           idx_dt = SurgeryDate,
                                           evt_dt = TransplantDate,
                                           end_dt = LastVisitDate,
                                           death_dt = DeathDate,
                                           append = TRUE,
                                           units = "months",
                                           adm_cnr_time = 24)

CIF <- estimate_cif_km(cmp_risk_data, evt = evt,evt_time = evt_time)
CIF_Sex <- estimate_cif_km(cmp_risk_data, evt = evt,evt_time = evt_time,group = Sex)

}
