% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{factor_desp}
\alias{factor_desp}
\title{factor_desp}
\usage{
factor_desp(
  df,
  group,
  includeNA = FALSE,
  round_to_100 = FALSE,
  drop.unused.levels = FALSE
)
}
\arguments{
\item{df}{Dataframe}

\item{group}{Optional grouping variable.}

\item{includeNA}{Logical; whether to include missing values as a level.}

\item{round_to_100}{Logical; whether rounded percentages should sum to 100.}

\item{drop.unused.levels}{Logical; whether to remove unused factor levels.}
}
\value{
a dataframe consisting of columns of character variables indicating the frequency and proportion of logical variables.
}
\description{
factor_desp
}
\details{
An internal function that calculate frequencies and proportion of levels in factor/categorical variables.
If there are more than 1 group, then fisher exact tests are applied to assess the between-group differences.
}
