% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_set_up.R
\name{factor_order}
\alias{factor_order}
\title{Order a factor variable by descending frequency}
\usage{
factor_order(var)
}
\arguments{
\item{var}{A factor or character vector to be reordered.}
}
\value{
A factor vector with levels sorted in descending order of frequency.
}
\description{
Order a factor variable by descending frequency
}
\details{
Automatically orders the levels of a factor variable by descending frequency, improving clarity in summary tables
}
\examples{


library(dplyr)

## Pulling out the checkbox questions (More than one may be applicable for the same header)
Comorbidities  <- cardio_data \%>\% select(Diabetes:CAD) \%>\% names()

## Cleaning the data by putting SurgeryType in descending order
## Converting the Comorbidities columns to checkbox and including a title.

work_d <- cardio_data \%>\%
 mutate(SurgeryType = factor_order(SurgeryType)) \%>\%
 check_box_convert(check_box_cols = Comorbidities,title = "Comorbidities¹")


table_one(df = work_d  ,
         group = Sex,
         caption =  "Summary table overall and stratified by sex",
         include_overall = "all",
         overall_label = "Overall",
         drop.unused.levels = FALSE,
         round_to_100 = TRUE,
         kable_output = TRUE,
         include_Missing = FALSE,
         print_unused = TRUE,
         Check_box = Comorbidities,
         Check_box_title = "Comorbidities¹")\%>\%
 kableExtra::footnote(
   general = "¹Patients could present with more than one comorbidity, totals may not sum to 100\%.",
   general_title = "",
   footnote_as_chunk = TRUE
 )
}
