% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format_pvalue}
\alias{format_pvalue}
\title{format_pvalue}
\usage{
format_pvalue(
  x,
  eps = 0.001,
  trim = TRUE,
  droptrailing0 = FALSE,
  pad = FALSE,
  ...
)
}
\arguments{
\item{x}{Numeric variable}

\item{eps}{Threshold used for displaying small p-values.}

\item{trim}{Logical; whether to trim leading whitespace.}

\item{droptrailing0}{Logical; whether to drop trailing zeroes.}

\item{pad}{Logical; whether to pad formatted values.}

\item{...}{Additional arguments passed to \code{format.pval()}.}
}
\value{
character variables reporting p-values
}
\description{
format_pvalue
}
\details{
An internal function that formats p-values according to the statistical guidelines of the Annals of Medicine.
}
