% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{k_sample_test}
\alias{k_sample_test}
\title{k_sample_test}
\usage{
k_sample_test(df, group)
}
\arguments{
\item{df}{Data frame containing the numeric variables.}

\item{group}{Factor grouping variable.}
}
\value{
a 2-tuple vectors reporting p-values
}
\description{
k_sample_test
}
\details{
An internal function that tests equality of location parameter using one way ANOVA with unequal
variance when mean is reported and nonparametric Kruskal-Wallis tests otherwise in comparing >2 groups.
}
