% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{numeric_desp}
\alias{numeric_desp}
\title{numeric_desp}
\usage{
numeric_desp(df, group)
}
\arguments{
\item{df}{Dataframe}

\item{group}{Optional grouping variable.}
}
\value{
a dataframe consisting of columns of character variables.
}
\description{
numeric_desp
}
\details{
An internal function that report mean, standard deviation, median and interquartile range by group.
}
