% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepare_survfit}
\alias{prepare_survfit}
\title{prepare survfit for survival table}
\usage{
prepare_survfit(surv_obj)
}
\arguments{
\item{surv_obj}{A \code{survfit} object to prepare for tabulation.}
}
\description{
prepare survfit for survival table
}
