% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{recode_missing}
\alias{recode_missing}
\title{recode_missing}
\usage{
recode_missing(x, na.value = NULL)
}
\arguments{
\item{x}{A vector in which coded missing values should be replaced.}

\item{na.value}{Values to replace with \code{NA}.}
}
\value{
input variable with NA
}
\description{
recode_missing
}
\details{
An internal function that replace missing value code with NA.
}
