% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{run_gray_test}
\alias{run_gray_test}
\title{Run Grays Test}
\usage{
run_gray_test(surv_obj, evt_type = 1:2)
}
\arguments{
\item{surv_obj}{A \code{survfit} object, such as one returned by \code{estimate_cif_km()}, containing grouped survival curves.}

\item{evt_type}{The numeric vector of interest. Default is 1:2}
}
\value{
A numeric value representing the p-value from Grays test.
}
\description{
Performs Grays test to compare survival distributions between two or more groups.
}
