% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{run_logrank_test}
\alias{run_logrank_test}
\title{Run Log-Rank Test for Survival Differences}
\usage{
run_logrank_test(surv_obj)
}
\arguments{
\item{surv_obj}{A \code{survfit} object, such as one returned by \code{estimate_km()}, containing grouped survival curves.}
}
\value{
A numeric value representing the p-value from the log-rank test.
}
\description{
Performs a log-rank test to compare survival distributions between two or more groups.
}
