% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_functions.R
\name{summarize_km}
\alias{summarize_km}
\title{Summarize Kaplan-Meier Estimates}
\usage{
summarize_km(
  fit,
  times = NULL,
  failure_fun = FALSE,
  kable_output = TRUE,
  caption = NULL,
  full_width = NULL,
  time_lab = "Times",
  overall_label = "Overall"
)
}
\arguments{
\item{fit}{A fitted survival object of class \code{survfit}.}

\item{times}{Optional numeric vector of time points at which to summarize the KM estimates. If \code{NULL}, uses \code{pretty(fit$time)}.}

\item{failure_fun}{Logical; if \code{TRUE}, returns failure probabilities (1 - survival).}

\item{kable_output}{Logical; if \code{TRUE}, formats the output using \code{kableExtra::kbl()}.}

\item{caption}{Optional character string for the table caption (used if \code{kable_output = TRUE}).}

\item{full_width}{Logical; passed to \code{kableExtra::kable_styling()} to control table width.}

\item{time_lab}{Character string; label to use for the time column in the output table.}

\item{overall_label}{Character string to label the overall summary column. Default is \code{"Overall"}.}
}
\value{
A data frame or a formatted \code{kableExtra} table summarizing survival or failure probabilities with confidence intervals.
}
\description{
Summarizes a fitted Kaplan-Meier survival object at specified time points, optionally transforming to failure probabilities,
and formats the output as a table with confidence intervals. See also \code{\link{construct_surv_cmprisk_var}},\code{\link{estimate_cif_km}}, \code{\link{summarize_cif}}, \code{\link{show_surv}}, and \code{\link{show_cif}}   for related functions.
}
\details{
The function summarizes the fitted Kaplan-Meier survival estimates at user-specified time points. If the model includes strata,
the output is stratified accordingly. Confidence intervals are included, and results are formatted as percentages with one decimal place.
If \code{failure_fun = TRUE}, the function returns failure probabilities instead of survival.
}
\examples{

survival_data <-  construct_surv_cmprisk_var(cardio_data,
                                            patid = PatientID,
                                            idx_dt = SurgeryDate,
                                            evt_dt = DeathDate,
                                            end_dt = LastVisitDate,
                                            append = TRUE,
                                            units = "months",
                                            adm_cnr_time = 24)

KM <- estimate_cif_km(survival_data, evt = evt,evt_time = evt_time)
summarize_km(KM,overall_label = "All patients",
time_lab = "Time since surgery (months)",caption = "Overall Survival by sex")

KM_Sex <- estimate_cif_km(survival_data, evt = evt,evt_time = evt_time,group = Sex)
summarize_km(KM_Sex,overall_label = "All patients",
time_lab = "Time since surgery (months)",caption = "Overall Survival for all patients")

}
