% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{summarize_mi_coxph}
\alias{summarize_mi_coxph}
\title{summarize_mi_coxph}
\usage{
summarize_mi_coxph(
  cox_mira,
  exponentiate = TRUE,
  alpha = 0.05,
  pval = c("omnibus", "both")
)
}
\arguments{
\item{cox_mira}{A \code{mira} object containing multiply imputed Cox model fits.}

\item{exponentiate}{Logical; whether to exponentiate estimates and confidence limits.}

\item{alpha}{Significance level used to calculate confidence intervals.}

\item{pval}{Character string specifying whether to report omnibus p-values
or both omnibus and coefficient-level p-values. Defaults to "omnibus".}
}
\description{
Summarize a Cox model fitted across multiply imputed data sets.
}
