% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{table_one_overall}
\alias{table_one_overall}
\title{table_one_overall}
\usage{
table_one_overall(
  df,
  total = TRUE,
  round_to_100 = FALSE,
  drop.unused.levels = FALSE,
  overall_label = "Overall"
)
}
\description{
table one with no stratifying variable
}
\keyword{internal}
