% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{table_one_stratify}
\alias{table_one_stratify}
\title{table_one_stratify}
\usage{
table_one_stratify(
  df,
  group,
  total = TRUE,
  round_to_100 = FALSE,
  drop.unused.levels = FALSE
)
}
\description{
table one with a stratifying variable
}
\keyword{internal}
