% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{two_sample_test}
\alias{two_sample_test}
\title{two_sample_test}
\usage{
two_sample_test(df, group)
}
\arguments{
\item{df}{Data frame containing the numeric variables.}

\item{group}{Factor grouping variable.}
}
\value{
a 2-tuple vectors reporting p-values
}
\description{
two_sample_test
}
\details{
An internal function that tests equality of location parameter using two sample t-tests with unequal
variance when mean is reported and nonparametric rank-sum tests otherwise in comparing of two groups.
}
