% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{updateWorksheet}
\alias{updateWorksheet}
\title{updateWorksheet}
\usage{
updateWorksheet(wb, sheetName, x, ...)
}
\arguments{
\item{wb}{a \code{wb} object}

\item{sheetName}{a name of the sheet to be updated}

\item{x}{a dataframe to be write in the \code{wb} object}

\item{...}{Additional arguments reserved for compatibility.}
}
\value{
a \code{wb} object
}
\description{
updateWorksheet
}
\details{
An internal function that adds a new worksheet or updates (remove and add) the existing worksheet in wb object.
}
