% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{add_atrisk}
\alias{add_atrisk}
\title{Create atrisk table}
\usage{
add_atrisk(
  p,
  surv_obj,
  space = -0.15,
  x_break = NULL,
  atrisk_init_pos = NULL,
  plot_theme = NULL
)
}
\description{
Create atrisk table
}
