% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_functions.R
\name{construct_surv_cmprisk_var}
\alias{construct_surv_cmprisk_var}
\title{construct survival or competing risks process}
\usage{
construct_surv_cmprisk_var(
  df,
  patid,
  idx_dt,
  evt_dt,
  end_dt,
  cmprisk = NULL,
  surv_varname = NULL,
  units = "days",
  append = FALSE,
  adm_cnr_time = NULL,
  ...
)
}
\arguments{
\item{df}{input data}

\item{patid}{the variable indicating subject/patient id}

\item{idx_dt}{the index date}

\item{evt_dt}{the date of the event occurrence. Its value should be NA for non-event subjects.}

\item{end_dt}{the date of the last follow-up}

\item{cmprisk}{Logical; indicates whether to create variables for competing risks. Defaults to \code{FALSE} when no competing event is included, and \code{TRUE} when a competing event is present.}

\item{surv_varname}{an option of character vector of length 2, the 1st of which is the name of the time variable; the other is the name of the event indicator.}

\item{units}{Character string specifying the unit of time for the time-to-event analysis. Accepted values are: "days (default) ", "weeks", "months", or "years".}

\item{adm_cnr_time}{a numeric vector specifying the time point at which administrative censoring is applied (in the same units as units).}

\item{...}{all competing event dates}
}
\value{
A data frame with patid, evt_time and evt.
}
\description{
The function creates time-to-event variables for survival or competing risks process.
}
\examples{

library(dplyr)

## For survival
construct_surv_cmprisk_var(cardio_data,
                          patid = PatientID,
                          idx_dt = SurgeryDate,
                          evt_dt = DeathDate,
                          end_dt = LastVisitDate,,
                          append = TRUE)

# For competing risks
construct_surv_cmprisk_var(cardio_data,
                          patid = PatientID,
                          idx_dt = SurgeryDate,
                          evt_dt = TransplantDate,
                          end_dt = LastVisitDate,
                          other_dt = DeathDate,
                          append = TRUE)

}
