% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{factor_desp}
\alias{factor_desp}
\title{titles_non_missing}
\usage{
factor_desp(
  df,
  group,
  includeNA = FALSE,
  round_to_100 = FALSE,
  drop.unused.levels = FALSE
)
}
\arguments{
\item{df}{Dataframe}
}
\value{
a dataframe consisting of columns of character variables indicating the frequency and proportion of logical variables.
}
\description{
Add a title to a dataframe
}
\details{
An internal function that calculate frequencies and proportion of levels in factor/categorical variables.
If there are more than 1 group, then fisher exact tests are applied to assess the between-group differences.
}
