% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{factor_order}
\alias{factor_order}
\title{factor_order}
\usage{
factor_order(var)
}
\arguments{
\item{var}{A factor or character vector to be reordered.}
}
\value{
A factor vector with levels sorted in descending order of frequency.
}
\description{
factor_order
}
\details{
Automatically orders the levels of a factor variable by descending frequency, improving clarity in summary tables
}
