% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_set_up.R
\name{factor_order}
\alias{factor_order}
\title{Order a factor variable by descending frequency.}
\usage{
factor_order(var)
}
\arguments{
\item{var}{A factor or character vector to be reordered.}
}
\value{
A factor vector with levels sorted in descending order of frequency.
}
\description{
Order a factor variable by descending frequency. To be used with \code{\link{table_one}}
}
\details{
Automatically orders the levels of a factor variable by descending frequency, improving clarity in summary tables
}
\examples{


library(dplyr)

cardio_data \%>\% select(SurgeryType) \%>\% table_one()

cardio_data \%>\%
 mutate(SurgeryType = factor_order(SurgeryType)) \%>\% select(SurgeryType) \%>\% table_one()

}
