% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one.R
\name{kable_table_one}
\alias{kable_table_one}
\title{kable_table_one}
\usage{
kable_table_one(
  tableone,
  caption = NULL,
  bold_variables = TRUE,
  full_width = TRUE,
  print_test = FALSE
)
}
\description{
creates a kable table one for rmarkdown reports
}
\keyword{internal}
