% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{kable_table_one}
\alias{kable_table_one}
\title{kable_table_one}
\usage{
kable_table_one(
  out,
  pval,
  include_Missing,
  total,
  print_test,
  caption,
  bold_variables
)
}
\description{
creates a kable table one for rmarkdown reports
}
\keyword{internal}
