% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event_functions.R
\name{summarize_cif}
\alias{summarize_cif}
\title{Summarize Cumulative Incidence Function (CIF)}
\usage{
summarize_cif(
  fit,
  times = NULL,
  kable_output = TRUE,
  caption = NULL,
  full_width = NULL,
  time_lab = "Time",
  evt_type = NULL,
  labels = c(`(s0)` = "Event free", `1` = "Event", `2` = "Competing", `3` =
    "Second Competing"),
  overall_label = "Overall"
)
}
\arguments{
\item{fit}{A fitted survival object of class \code{survfit}.}

\item{times}{Optional numeric vector of time points at which to summarize the CIF estimates. If \code{NULL}, uses \code{pretty(fit$time)}.}

\item{kable_output}{Logical; if \code{TRUE}, formats the output using \code{kableExtra::kbl()}.}

\item{caption}{Optional character string for the table caption (used if \code{kable_output = TRUE}).}

\item{full_width}{Logical; passed to \code{kableExtra::kable_styling()} to control table width.}

\item{time_lab}{Character string; label to use for the time column in the output table.}

\item{evt_type}{Numeric; Indicate which events will be outputted in the kable table; default is all events}

\item{labels}{Character string; labels for the events in the kable table when include_event_type is TRUE. Default is c('(s0)' = "Event free",'1' = "Event",'2'="Competing",'3' = "Other")}

\item{overall_label}{Character string to label the overall summary column. Default is \code{"Overall"}.}
}
\value{
A data frame or a formatted \code{kableExtra} table summarizing survival or failure probabilities with confidence intervals.
}
\description{
Summarizes a fitted Cumulative Incidence Function (CIF) survival object at specified time points, optionally transforming to failure probabilities,
and formats the output as a table with confidence intervals.
}
\details{
The function summarizes the fitted Kaplan-Meier survival estimates at user-specified time points. If the model includes strata,
the output is stratified accordingly. Confidence intervals are included, and results are formatted as percentages with one decimal place.
}
\examples{

library(dplyr)
## Example without a covariate

construct_surv_cmprisk_var(cardio_data,
                          patid = PatientID,
                          idx_dt = SurgeryDate,
                          evt_dt = TransplantDate,
                          end_dt = LastVisitDate,
                          append = TRUE,
                          cmprisk = TRUE,
                          other = DeathDate,
                          units = "months") \%>\%
 estimate_cif_km(evt = evt,evt_time = evt_time) \%>\%
 summarize_cif(evt_type = c(0,1,2),
               labels = c('(s0)' = "Event free",'1' = "Transplant",'2'="Death"),time_lab = "Time since surgery (months)")


## Example with a covariate

construct_surv_cmprisk_var(cardio_data,
                          patid = PatientID,
                          idx_dt = SurgeryDate,
                          evt_dt = TransplantDate,
                          end_dt = LastVisitDate,
                         append = TRUE,
                          cmprisk = TRUE,
                          other = DeathDate,
                          units = "months") \%>\%
 estimate_cif_km(evt = evt,evt_time = evt_time,Sex) \%>\%
 summarize_cif(evt_type = c(1,2),
               labels = c('(s0)' = "Event free",'1' = "Transplant",'2'="Death"),time_lab = "Time since surgery (months)")

}
