% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desp_table.R
\name{table_one}
\alias{table_one}
\title{table_one}
\usage{
table_one(
  df,
  group,
  datadic = NULL,
  var_name,
  var_desp,
  seed = 123,
  include_overall = c("none", "group", "all"),
  total = TRUE,
  pval = TRUE,
  continuous = c("mediqr", "meansd"),
  kable_output = TRUE,
  caption
)
}
\arguments{
\item{df}{A data frame consisting of numeric, logical, and factor variables with or without a grouping variable}

\item{group}{Name of the grouping variable.}

\item{datadic}{A data frame containing a data dictionary of variable names and their descriptions.}

\item{var_name}{the column name of `data_dict` that
contains the variable names. Only required if the column name is not "var_name"}

\item{var_desp}{the column name of `data_dict` that
contains the variable descriptions Only required if the column name is not "var_desp"}

\item{seed}{Sets a seed}

\item{include_overall}{Character string specifying whether and how to include an overall summary.}

\item{total}{Logical variariable to report total N. Default is TRUE
Must be one of:
\itemize{
  \item `"none"`: Do not include an overall summary.
  \item `"group"`: Include an overall summary only for observations with non-missing values in the grouping variable.
  \item `"all"`: Include an overall summary for all observations, regardless of missingness in the grouping variable.
}
Default is `"none"`.}

\item{pval}{Option to report p-value}

\item{continuous}{Select which continuous descriptors to include using "mediqr" for the median with interquartile range and "meansd" for the mean and standard deviation}

\item{kable_output}{Outputs table as a formatted kable table and includes the column var_desp, N, Stat and p-value}
}
\value{
The function returns a dataframe, rows of which are summary statistics depending on the variable types.
}
\description{
Creates a table of summary statistics
}
\details{
\code{table_one} calculate the selected summary statistics for continuous, logical,
and factor variables per statistical guidelines of the Annals of medicine. If a group variable is provided, then
it will also assess the between-group difference.The input data frame should only consists of numeric, logical
and factor variables. Factor variables with
only two levels should be converted to logical variables. Date and datetime variables should be removed.
}
\examples{
set.seed(0)
df<- data_frame(sex   = factor(c(rep("F", 90), rep("M", 900))),
                grade = factor(sample(c("A", "B", "C"), 990, replace= TRUE), c("A", "B", "C", "D")),
                income=  100 * (rnorm(990) + 5),
                dm= rbernoulli(990, p= .5),
                af= rbernoulli(990, p= .95)) \%>\%
  mutate(weight= if_else( sex=="F" & income>500, 3, 1),
         income= ifelse(income<456, NA, income),
         sex   = ifelse(runif(990)<.2, NA, sex),
         sex   = factor(sex, 1:2, labels = c("Female", "Male")),
         grade = ifelse(runif(990)<.25, NA, grade),
         grade   = factor(grade, 1:4, labels = c("A", "B", "C", "D")))

 datadic<- data.frame(var_name= c("sex", "grade", "income", "dm", "af"),
                      var_desp= c("Sex", "Grade", "Household income",
                                  "Presence of diabetes mellitus", "African American"),
                     stringsAsFactors = FALSE)

table_one(df, sex)
table_one(df, sex, datadic= datadic)
}
