% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desp_table.R
\name{table_one}
\alias{table_one}
\title{table_one}
\usage{
table_one(
  df,
  group,
  datadic = NULL,
  var_name,
  var_desp,
  seed = 123,
  include_overall = c("none", "group", "all"),
  total = TRUE,
  pval = TRUE,
  print_test = FALSE,
  continuous = "mediqr",
  round_to_100 = FALSE,
  drop.unused.levels = FALSE,
  kable_output = TRUE,
  caption = NULL,
  overall_label = "Overall"
)
}
\arguments{
\item{df}{A data frame consisting of numeric, logical, and factor variables with or without a grouping variable}

\item{group}{Name of the grouping variable (optional).}

\item{datadic}{datadic A data frame serving as a data dictionary, containing variable names and their descriptions.}

\item{var_name}{the column name of `data_dict` that
contains the variable names. Only required if the column name is not "var_name"}

\item{var_desp}{the column name of `data_dict` that
contains the variable descriptions Only required if the column name is not "var_desp"}

\item{seed}{An optional seed value for reproducibility of p-values}

\item{include_overall}{Character string specifying whether and how to include an overall summary.
Must be one of:
\itemize{
  \item `"none"`: Do not include an overall summary (default).
  \item `"group"`: Include an overall summary only for observations with non-missing values in the grouping variable.
  \item `"all"`: Include an overall summary for all observations, regardless of missingness in the grouping variable.
}
Default is `"none"`.}

\item{total}{Logical; whether to report the total N. Default is `TRUE`.}

\item{pval}{Logical; whether to report p-values for between-group comparisons. Default is `TRUE`.}

\item{print_test}{Logical. If TRUE, the output will include the type of statistical test applied to each variable. Default is `FALSE`.}

\item{continuous}{Character string specifying the summary statistics for continuous variables.
Must be one of:
 \itemize{
   \item `"mediqr"`: Median and interquartile range.
   \item `"meansd"`: Mean and standard deviation.
   \item `"c("mediqr","meansd")"`: Both median/IQR and mean/SD.
 }}

\item{round_to_100}{force rounded total to add up to 100% calculated using the largest remainder method for factor variables}

\item{drop.unused.levels}{Removes factor variables with 0. Levels with zero are not included in statistical tests}

\item{kable_output}{Logical; if `TRUE`, outputs a formatted `kable` table including variable descriptions, N, statistics, and p-values.}
}
\value{
A data frame containing summary statistics by variable type, optionally stratified by group and formatted for reporting, or a formatted kable table if `kable_output = TRUE`..
}
\description{
Generates a table of summary statistics for descriptive analysis.
}
\details{
The `table_one` function computes summary statistics for continuous, logical, and factor variables,
following the statistical reporting guidelines of the *Annals of Medicine*. If a grouping variable is provided,
the function can also evaluates between-group differences. The input data frame should consist only of numeric,
logical, and factor variables. Factor variables with only two levels should be converted to logical variables.
Date and datetime variables should be excluded.
}
\examples{
set.seed(0)
df<- data_frame(sex   = factor(c(rep("F", 90), rep("M", 900))),
                grade = factor(sample(c("A", "B", "C"), 990, replace= TRUE), c("A", "B", "C", "D")),
                income=  100 * (rnorm(990) + 5),
                dm= rbernoulli(990, p= .5),
                af= rbernoulli(990, p= .95)) \%>\%
  mutate(weight= if_else( sex=="F" & income>500, 3, 1),
         income= ifelse(income<456, NA, income),
         sex   = ifelse(runif(990)<.2, NA, sex),
         sex   = factor(sex, 1:2, labels = c("Female", "Male")),
         grade = ifelse(runif(990)<.25, NA, grade),
         grade   = factor(grade, 1:4, labels = c("A", "B", "C", "D")))

 datadic<- data.frame(var_name= c("sex", "grade", "income", "dm", "af"),
                      var_desp= c("Sex", "Grade", "Household income",
                                  "Presence of diabetes mellitus", "African American"),
                     stringsAsFactors = FALSE)

table_one(df, sex)
table_one(df, sex, datadic= datadic)
}
